fix(pallets): resolve all clippy warnings across pallets and runtime
- Fix excessive_nesting warnings by adding targeted allow attributes
  to Substrate mutate/iter patterns where refactoring isn't practical
- Fix redundant_closure warnings by using function references directly
- Fix unnecessary_map_or by using is_none_or method
- Fix clone_on_copy by removing clone calls on Copy types
- Fix unnecessary_cast in octopus pallet
- Update runtime imports: remove unused imports and replace deprecated
  create_runtime_str! macro with Cow::Borrowed

Affected pallets: governance, storage, semantic, autonomous, boomerang,
device, dispute, epoch, lifecycle, octopus, validator

+++ pallets/octopus/src/lib.rs
@@ -626,7 +626,7 @@ pub mod pallet {
 
         fn calculate_target_subnodes(throughput: Perbill) -> u32 {
             let pct = throughput.deconstruct() / 10_000_000;
-            let scaled = (pct as u32).saturating_mul(10);
+            let scaled = pct.saturating_mul(10);
             let divisor = 225u32;
             let result = scaled.saturating_add(divisor.saturating_sub(1)) / divisor;
 
@@ -636,6 +636,7 @@ pub mod pallet {
             result.clamp(min, max)
         }
 
+        #[allow(clippy::excessive_nesting)]
         fn process_deactivations(block_number: BlockNumberFor<T>) {
             let duration = T::DeactivationDurationBlocks::get();
